Move to more appropiate location
