+++ keanu-project/src/test/java/io/improbable/keanu/algorithms/graphtraversal/DifferentiableCheckerTest.java
@@ -1,7 +1,9 @@
-package io.improbable.keanu.vertices.dbl;
+package io.improbable.keanu.algorithms.graphtraversal;
 
+import io.improbable.keanu.algorithms.graphtraversal.DifferentiableChecker;
 import io.improbable.keanu.network.BayesianNetwork;
 import io.improbable.keanu.vertices.bool.probabilistic.BernoulliVertex;
+import io.improbable.keanu.vertices.dbl.DoubleVertex;
 import io.improbable.keanu.vertices.dbl.nonprobabilistic.ConstantDoubleVertex;
 import io.improbable.keanu.vertices.dbl.nonprobabilistic.operators.unary.FloorVertex;
 import io.improbable.keanu.vertices.dbl.probabilistic.GaussianVertex;
